+++ flutter_inappwebview_forge/lib/src/in_app_webview/in_app_webview_controller.dart
@@ -301,6 +301,12 @@ class InAppWebViewController {
   Future<void> setSettings({required InAppWebViewSettings settings}) =>
       platform.setSettings(settings: settings);
 
+  ///{@macro flutter_inappwebview_forge_platform_interface.PlatformInAppWebViewController.setBackgroundColor}
+  ///
+  ///{@macro flutter_inappwebview_forge_platform_interface.PlatformInAppWebViewController.setBackgroundColor.supported_platforms}
+  Future<void> setBackgroundColor(Color color) =>
+      platform.setBackgroundColor(color);
+
   ///{@macro flutter_inappwebview_forge_platform_interface.PlatformInAppWebViewController.getSettings}
   ///
   ///{@macro flutter_inappwebview_forge_platform_interface.PlatformInAppWebViewController.getSettings.supported_platforms}