+++ flutter_inappwebview_forge_android/android/src/main/kotlin/com/emirkanacar/flutter_inappwebview_forge_android/webview/WebViewChannelDelegateMethods.kt
@@ -23,6 +23,7 @@ enum class WebViewChannelDelegateMethods {
     isLoading,
     takeScreenshot,
     setSettings,
+    setBackgroundColor,
     getSettings,
     close,
     show,